Move generic workflow enums into engine-owned module
Relocate WorkflowType, BenchmarkTaskType, ReportCheckTypes,
ModelStatusTypes, EvalLimitMode, AgenticTracesMode, VersionMode and
ModelType from workflows/workflow_types.py to a new engine-owned
workflow_module/engine_types.py, and update the 13 engine import sites
to use it.

workflows/workflow_types.py keeps the Tenstorrent-specific taxonomies
(DeviceTypes, SystemTopology, InferenceEngine, ModelSource,
WorkflowVenvType) and re-exports the moved enums, so existing v1
callers are untouched. This establishes the dependency direction the
standalone-framework extraction needs: the adapter layer may depend on
the engine core, never the reverse. Purely mechanical; enum member
order preserved so IntEnum values are unchanged.
