+++ test_module/llm_tests/llm_eval_tests.py
@@ -21,8 +21,8 @@ from llm_module.eval_configs import get_llm_eval_tasks
 from reference_config.evals.eval_config import accept_eval_score, resolve_eval_reference
 from report_module.schema import Block
 from workflow_module import accept_blocks
+from workflow_module.engine_types import EvalLimitMode
 from workflows.utils import run_command
-from workflows.workflow_types import EvalLimitMode
 
 from .._test_common import ReportCheckTypes, TestStatus, block_id
 from ..context import MediaContext
